+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/populator/BaseBlockPopulator.java
@@ -1,15 +1,19 @@
 package com.elmakers.mine.bukkit.world.populator;
 
 import java.util.Random;
+import java.util.Set;
 
 import org.bukkit.Chunk;
 import org.bukkit.Material;
 import org.bukkit.World;
+import org.bukkit.block.Biome;
 import org.bukkit.block.Block;
 import org.bukkit.configuration.ConfigurationSection;
 
 import com.elmakers.mine.bukkit.magic.MagicController;
+import com.elmakers.mine.bukkit.utility.ConfigurationUtils;
 import com.elmakers.mine.bukkit.world.BlockResult;
+import com.elmakers.mine.bukkit.world.spawn.SpawnResult;
 
 public abstract class BaseBlockPopulator extends MagicChunkPopulator {
     private int maxY = 255;
@@ -17,6 +21,8 @@ public abstract class BaseBlockPopulator extends MagicChunkPopulator {
     private int maxAirY = 255;
     private int cooldown;
     private long lastPopulate;
+    private Set<Biome> biomes;
+    private Set<Biome> notBiomes;
 
     @Override
     public boolean load(ConfigurationSection config, MagicController controller) {
@@ -27,6 +33,8 @@ public abstract class BaseBlockPopulator extends MagicChunkPopulator {
         minY = config.getInt("min_y", minY);
         maxAirY = config.getInt("max_air_y", maxAirY);
         cooldown = config.getInt("cooldown", 0);
+        biomes = ConfigurationUtils.loadBiomes(ConfigurationUtils.getStringList(config, "biomes"), controller.getLogger(), "block populator");
+        notBiomes = ConfigurationUtils.loadBiomes(ConfigurationUtils.getStringList(config, "not_biomes"), controller.getLogger(), "block populator");
         return true;
     }
 
@@ -44,6 +52,10 @@ public abstract class BaseBlockPopulator extends MagicChunkPopulator {
                         if (now < lastPopulate + cooldown) continue;
                         lastPopulate = now;
                     }
+                    if (biomes != null && !biomes.contains(block.getBiome()))
+                        continue;
+                    if (notBiomes != null && notBiomes.contains(block.getBiome()))
+                        continue;
 
                     populate(block, random);
                 }